added endpoint for fetching full info about rehearsal. closes #215

+++ app/Http/Controllers/Management/RehearsalsController.php
@@ -6,8 +6,8 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Management\RehearsalsFilterManagementRequest;
 use App\Http\Requests\Management\UpdateRehearsalRequest;
 use App\Http\Resources\Management\RehearsalDetailedResource;
+use App\Http\Resources\Users\RehearsalResource;
 use App\Models\Rehearsal;
-use Exception;
 use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
@@ -21,8 +21,6 @@ use Illuminate\Http\Response;
 class RehearsalsController extends Controller
 {
     /**
-     * @param  RehearsalsFilterManagementRequest  $filterRequest
-     * @return AnonymousResourceCollection
      * @throws AuthorizationException
      */
     public function index(RehearsalsFilterManagementRequest $filterRequest): AnonymousResourceCollection
@@ -31,32 +29,31 @@ class RehearsalsController extends Controller
 
         $rehearsals = Rehearsal::query()
             ->filter($filterRequest)
-            ->with(['user', 'band'])
             ->orderBy('id')
             ->get();
 
-        return RehearsalDetailedResource::collection($rehearsals);
+        return RehearsalResource::collection($rehearsals);
     }
 
     /**
-     * @param  Rehearsal  $rehearsal
-     * @param  UpdateRehearsalRequest  $request
-     * @return RehearsalDetailedResource
+     * @throws AuthorizationException
      */
     public function update(Rehearsal $rehearsal, UpdateRehearsalRequest $request): RehearsalDetailedResource
     {
+        $this->authorize('manage', $rehearsal);
+
         $rehearsal->update($request->getStatusAttribute());
 
         return new RehearsalDetailedResource($rehearsal);
     }
 
     /**
-     * @param  Rehearsal  $rehearsal
-     * @return JsonResponse
-     * @throws Exception
+     * @throws AuthorizationException
      */
     public function delete(Rehearsal $rehearsal): JsonResponse
     {
+        $this->authorize('manage', $rehearsal);
+
         $rehearsal->delete();
 
         return response()->json(null, Response::HTTP_NO_CONTENT);
